requirements.txt
fastapi
openai
python-docx
uvicorn
